Fix hw_tests build
