+++ pigglet/tests/tcp.rs
@@ -7,9 +7,11 @@ use pigdef::config::HardwareConfig;
 use pigdef::config::HardwareConfigMessage::{GetConfig, NewConfig, NewPinConfig};
 use pigdef::description::HardwareDescription;
 use pigdef::pin_function::PinFunction::Output;
+use piggpio::config::CONFIG_FILENAME;
 use pignet::tcp_host;
 use serial_test::serial;
 use std::net::IpAddr;
+use std::path::PathBuf;
 use std::time::Duration;
 
 #[path = "../../piggui/tests/support.rs"]
@@ -57,6 +59,18 @@ async fn reconnect_tcp() {
     .await;
 
     pass(&mut pigglet);
+}
+
+#[allow(dead_code)]
+pub fn delete_configs() {
+    let crate_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+    let workspace_dir = crate_dir.parent().expect("Failed to get parent dir");
+    let config_file = workspace_dir.join(CONFIG_FILENAME);
+    println!("Deleting file: {config_file:?}");
+    let _ = std::fs::remove_file(config_file);
+    let config_file = workspace_dir.join("target/debug/").join(CONFIG_FILENAME);
+    println!("Deleting file: {config_file:?}");
+    let _ = std::fs::remove_file(config_file);
 }
 
 #[tokio::test]
