+++ piggui/Cargo.toml
@@ -32,6 +32,7 @@ usb = ["pigdef/usb", "pignet/usb"]
 [dependencies]
 pigdef = { path = "../pigdef", version = "0.7", default-features = false, features = ["std"] }
 pignet = { path = "../pignet", version = "0.7" }
+piggpio = { path = "../piggpio", version = "0.7" }
 
 chrono = { version = "0.4", default-features = false, features = ["now", "serde"] }
 anyhow = { version = "1", default-features = false, features = ["std"] }
